Spooky/app_auth/serializers.py
# auth/serializers.py
from rest_framework import serializers
from django.contrib.auth.models import User
from django_otp.plugins.otp_totp.models import TOTPDevice
from .models import UserProfile



class UserProfileSerializer(serializers.ModelSerializer):
    class Meta:
        model = UserProfile
        fields = ['id', 'user', 'phone_number', 'profile_picture', 'reputation', 'level']
